refactor: review -- static only classes to methods
