+++ tests/core/services/test_opossum_file_writer.py
@@ -5,7 +5,7 @@
 from pathlib import Path
 from zipfile import ZipFile
 
-from opossum_lib.core.services.opossum_file_writer import OpossumFileWriter
+from opossum_lib.core.services.opossum_file_writer import write_opossum_file
 from opossum_lib.shared.constants import (
     INPUT_JSON_NAME,
     OUTPUT_JSON_NAME,
@@ -22,7 +22,7 @@ def test_only_input_information_available_writes_only_input_information(
     )
     output_path = tmp_path / "output.opossum"
 
-    OpossumFileWriter.write(opossum_file_content, output_path)
+    write_opossum_file(opossum_file_content, output_path)
 
     with ZipFile(output_path, "r") as zip_file:
         assert zip_file.namelist() == [INPUT_JSON_NAME]
@@ -34,7 +34,7 @@ def test_input_and_output_information_available_writes_both(
     opossum_file_content = opossum_file_faker.opossum_file_content()
     output_path = tmp_path / "output.opossum"
 
-    OpossumFileWriter.write(opossum_file_content, output_path)
+    write_opossum_file(opossum_file_content, output_path)
 
     with ZipFile(output_path, "r") as zip_file:
         assert INPUT_JSON_NAME in zip_file.namelist()
